requirements.txt
streamlit>=1.40.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-groq>=0.2.0
langchain-chroma>=0.1.2
langchain-huggingface>=0.1.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
edge-tts>=6.1.18
pypdf>=5.0.0
python-dotenv>=1.0.1
PyMuPDF>=1.24.0
pytesseract>=0.3.10
Pillow>=10.4.0